Remove unnecessary directories/flattne structure
