+++ core/src/main/scala/geotrellis/server/TmsReification.scala
@@ -1,0 +1,17 @@
+package geotrellis.server
+
+import geotrellis.vector.Extent
+import com.azavea.maml.ast.{Literal, MamlKind}
+import cats._
+import cats.data.EitherT
+import cats.effect._
+import simulacrum._
+
+import java.util.UUID
+
+
+@typeclass trait TmsReification[A] {
+  @op("kind") def kind(self: A): MamlKind
+  @op("tmsReification") def tmsReification(self: A, buffer: Int)(implicit contextShift: ContextShift[IO]): (Int, Int, Int) => IO[Literal]
+}
+
