--- core/src/main/scala/geotrellis/server/core/conf/LoadConf.scala
@@ -1,25 +1,0 @@
-package geotrellis.server.core.conf
-
-import cats.effect.IO
-import com.typesafe.config.ConfigFactory
-import pureconfig.error.ConfigReaderException
-
-import java.net.URI
-import scala.reflect.ClassTag
-
-
-object LoadConf {
-
-  import pureconfig._
-
-  def apply(configFile: String = "application.conf") = new {
-    def as[Conf: ClassTag: ConfigReader]: IO[Conf] =
-      IO {
-        loadConfig[Conf](ConfigFactory.load(configFile))
-      }.flatMap {
-        case Left(e) => IO.raiseError[Conf](new ConfigReaderException[Conf](e))
-        case Right(config) => IO.pure(config)
-      }
-  }
-}
-
