--- core/src/main/scala/geotrellis/server/core/package.scala
@@ -1,6 +1,0 @@
-package geotrellis.server
-
-
-package object core {
-}
-
