--- core/src/main/scala/geotrellis/server/core/wcs/Constants.scala
@@ -1,11 +1,0 @@
-package geotrellis.server.core.wcs
-
-
-object Constants {
-  // Map of incoming format string to normalized format string for supported formats.
-  val SUPPORTED_FORMATS =
-    Map(
-      "geotiff" -> "geotiff",
-      "geotif" -> "geotiff"
-    )
-}
