--- core/src/main/scala/geotrellis/server/core/wcs/ops/package.scala
@@ -1,9 +1,0 @@
-package geotrellis.server.core.wcs
-
-import geotrellis.spark._
-
-
-package object ops {
-  type MetadataCatalog = Map[String, (Seq[Int], Option[TileLayerMetadata[SpatialKey]])]
-}
-
