+++ core/src/main/scala/geotrellis/server/wcs/WcsService.scala
