+++ core/src/main/scala/geotrellis/server/wcs/ops/package.scala
@@ -1,0 +1,9 @@
+package geotrellis.server.wcs
+
+import geotrellis.spark._
+
+
+package object ops {
+  type MetadataCatalog = Map[String, (Seq[Int], Option[TileLayerMetadata[SpatialKey]])]
+}
+
